+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+# 1.0.6
+- Fix a crash on startup or during backup when backup data is corrupted. ([HeatherComputer/AdvancedBackups#110](https://github.com/HeatherComputer/AdvancedBackups/issues/110))
+
+* * *
+
 # 1.0.5
-- Fix a rare crash when entering a singleplayer world. ([HeatherComputer/AdvancedBackups#110](https://github.com/HeatherComputer/AdvancedBackups/issues/110))
+- Fix a rare crash when entering a singleplayer world. ([HeatherComputer/AdvancedBackups#87](https://github.com/HeatherComputer/AdvancedBackups/issues/87))
 - Fix a potential crash during backup when a backup directory is missing.
 - Fix a potential crash when cleaning up old backup files.